% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{package}
\name{icudt-package}
\alias{icudt-package}
\title{ICU Data Files}
\description{
t.b.d.
}
\author{
Marek Gagolewski \email{gagolews@rexamine.com}\cr
ICU4C was developed by IBM and others.
The Unicode Character Database is due to Unicode, Inc.;
see the COPYRIGHTS file for more details.
}
\references{
\emph{\pkg{stringi} Package homepage}, \url{http://stringi.rexamine.com}

\emph{ICU -- International Components for Unicode}, \url{http://www.icu-project.org/}

\emph{ICU4C API Documentation}, \url{http://www.icu-project.org/apiref/icu4c/}

\emph{The Unicode Consortium}, \url{http://www.unicode.org/}
}

